legacy/django-cra/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
xlrd
whitenoise
django-allauth
djangorestframework-simplejwt
requests